pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "IRM Project"
authors = ["Ada-lave <113159483+Ada-lave@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.6"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
